rename attributes

+++ migrations/20240401225526_create_appliance.up.sql
@@ -1,11 +1,11 @@
 CREATE TABLE applience
 (
   id                   BIGSERIAL NOT NULL PRIMARY KEY,
-  student_id           INTEGER NOT NULL,
-  department_id        INTEGER NOT NULL,
-  exam_subject_1_id    INTEGER NOT NULL,
-  exam_subject_2_id    INTEGER NOT NULL,
-  exam_subject_3_id    INTEGER NOT NULL,
+  student_id           BIGSERIAL NOT NULL,
+  department_id        BIGSERIAL NOT NULL,
+  exam_subject_1_id    BIGSERIAL NOT NULL,
+  exam_subject_2_id    BIGSERIAL NOT NULL,
+  exam_subject_3_id    BIGSERIAL NOT NULL,
   FOREIGN KEY (department_id) REFERENCES department(id),
   FOREIGN KEY (exam_subject_1_id) REFERENCES exam_subject(id),
   FOREIGN KEY (exam_subject_2_id) REFERENCES exam_subject(id),
